+++ src/main/java/org/wikiclean/languages/Portuguese.java
@@ -1,0 +1,20 @@
+package org.wikiclean.languages;
+
+import java.util.List;
+import java.util.regex.Pattern;
+
+public class Portuguese extends Language {
+  public Portuguese() {
+    super("pt");
+  }
+
+  @Override
+  protected List<Pattern> footerPatterns() {
+    return footerPatterns("Ver também", "Notas", "Referências", "Ligações externas");
+  }
+
+  @Override
+  protected List<Pattern> categoryLinkPatterns() {
+    return categoryLinkPatterns("Categoria");
+  }
+}