+++ HISTORY.md
@@ -4,7 +4,21 @@ TBD
 
 + Simplified adding new languages by extracting language-specific cleanup into separate class
   (contribution by [fabrichter](https://github.com/fabrichter))
-+ Added support for French
++ Added basic support for 13 languages:
+    * French
+    * Breton
+    * Catalan
+    * Greek
+    * Esperanto
+    * Spanish
+    * Galician
+    * Italian
+    * Dutch
+    * Polish
+    * Portuguese
+    * Russian
+    * Ukrainian
+
 
 Version 1.1
 ===========